Don't log headings without content
The netconf server collects data until an XML tag is completed before
pretty printing received data. Each time data is logged, a heading
like the following is printed:

= CT_NETCONFC ==== 28-Sep-2015::16:43:46,842 ===================================
= Client <0.194.0> <----- {"127.0.0.1",2060} ===================================

This commit removes printing of this header if there is no data to be
printed below - i.e. if the XML tag is not yet complete and we are
waiting for more data.