epp: Add the pre-defined macro OTP_RELEASE
Add a new pre-defined macro called OTP_RELEASE that will expand
to an integer being the OTP version. Thus, in OTP 19 the value will
be the integer 19.

The OTP_RELEASE macro is particularly useful in order to have
different source code depending on new language features or new
features in the type specification syntax. Those features are only
introduced in major versions of OTP.

To be truly useful, the -if preprocessor directive need to be
implemented. That is the purpose of the next commit.

Code that will need to work in both OTP 18 and OTP 19 can be
structured in the following way:

-ifdef(OTP_RELEASE).
%% Code that only works in OTP 19 and later.
-else.
%% Code that will work in OTP 18.
-endif.