beam_reorder: Eliminate compiler crash
c288ab87 added beam_reorder to move get_tuple_element instructions.

Compiling code such as the following would crash the compiler:

  alloc(_U1, _U2, R) ->
    V = R#alloc.version,
    Res = id(V),
    _ = id(0),
    Res.

The crash would occur because the following two instructions:

    {get_tuple_element,{x,2},1,{x,1}}.
    {allocate_zero,1,2}.

were swapped and rewritten to:

    {allocate_zero,1,1}.
    {get_tuple_element,{x,2},1,{x,1}}.

That transformation is not safe because the allocate_zero instruction
would kill {x,2}, which is the register that is holding the reference
to the tuple. Only do the transformation when the tuple reference is
in an x register with a lower number than the destination register.