Avoid always updating inet_dns in the primary bootstrap
The include file inet_dns_record_adts.hrl is generated by the
Perl script inet_dns_record_adts.pl in a non-deterministic way.
That is, every time the script is run, the functions will be in
a different order. That will cause inet_dns.beam in the primary
bootstrap to be updated every time the bootstrap is updated,
even though there is no actual code change.

Modify the Perl script to sort the keys pulled out from hashes
to make the order deterministic.