Remove useless 'catch' in start_it/1
The last clause in start_it/1 calls a function in some module.

It goes to great length to catch any exception and pass them on
unchanged, and if there was a normal return, it will just return
the return value. It can been seen that the entire 'catch'
construction with the reference trick is totally unnecessary.