Save error reasons for TLS distribution connections
When establishing an outbound connection for TLS distribution, let's
hold on to the failure reasons and use them as exit reasons.  These
exit reasons are normally invisible, but they can be seen in the logs
after calling net_kernel:verbose(1).

While there are trace messages in the code already, those require
recompiling the module with a special flag, which is more cumbersome
than changing the net_kernel verbosity level at run time.