TLS distribution: wait for code server
As described in the comments in the patch, doing a TLS handshake
requires the crypto module to be loaded.  The crypto module needs the
code server to find its NIF library.  However, there is a time window
between opening the listening ports for distribution and starting the
code server, and if we get an incoming connection in that time window,
the node would believe that it's alive, but it wouldn't actually
accept any more connections.