Add option to override private key in ssh connect
ssh:connect/3,4 gets the user's private key from either the default
directory (~/.ssh) of the user or from the directory specified in the
'user_dir' ssh option. This change introduces new ssh:connect/3,4
options that lets the user specify the private key at the API level
instead of reading it from the file system. This is particularly
useful when the keys are stored in a database.

TODO
 * Support for ECDSA keys (all variants)
 * Documentation update