sys_pre_expand: Clean up data structures
The handling of non-remote calls is messy, with several lookups
to determine whether the call is local or to some imported module.

We can simplify the code if we keep a map that immediately gives
us the answer. Here is an example of what the map entries look
like:

  {f,1} => local
  {foldl,3} => {imported,lists}

That is, there should be a local call to f/1 and a remote call
to lists:foldl/3.

Note that there is no longer any need to keep the set of all defined
functions in the state record.