jinterface: fix writing small compressed values
This is a regression of 4390e43558 in the OtpOutputStream class.

We can not call java.util.zip.DeflaterOutputStream.close() in the finally
block of the OtpOutputStream.write_compressed(OtpErlangObject, int) method.
This leads to a NullPointerException when encoding "{}" which is caused by
the DeflaterOutputStream trying to write bytes to the deflater which was
"destroyed" by calling java.util.zip.Deflater.end().
Further possibilities to call close() in the finally block are not suitable
either (see the comment in the source). This leaves no choice but to revert
the change from 4390e43558 in this class (and add an appropriate test case).