erts: Fix bug in setnode/2
that could lead strange things to happen when renaming a node with
a name that already exist in node and dist tables.

Problem:
erts_set_this_node() is a bit brutal and does not handle the case
when an old remote node name is reused as the new local node name.

Solution:
Treat erts_this_node and erts_this_dist_entry as all the others
with correct reference counting.