Support SSH key callback module options
This patch allows extra callback options to be passed to the module
implementing the SSH callback module behaviour.

A module implementing the SSH key callback API is used to customize
the handling of public key. This patch allows extra callback options
to be passed to the module implementing the SSH callback module
behaviour.

The key_cb option has been changed:

  {key_cb, atom()} -> {key_cb, key_cb()}

  Where:
	key_cb() :: atom() | {atom(), list()}

The callback options, if specified, is made available to the callback
module via the options passed to it under the key 'key_cb_private'.

More details and some backgorund is available here[1].

[1]: http://erlang.org/pipermail/erlang-patches/2015-November/004800.html