prim_file: Suppress a dialyzer warning
Kostis Sagonas pointed out that there is a dialyzer warning
for constructing an improper list in the following clause:

translate_response(?FILE_RESP_N2DATA = X,
		   [<<_:64, _:64, _:64>> | <<>>] = Data) ->
    {error, {bad_response_from_port, [X | Data]}};

I don't want to change the code to somehow eliminate the warning. An
improper list has already been constructed in the efile driver itself,
and that would be difficult to fix.

Therefore, tell dialyzer to ignore warnings for improper lists
in translate_response/2.