ssh: AEAD_AES_(128|256)_GCM and aes(128|256)-gcm@openssh.com crypto
Note that the rfc5647 is ambigous so this implementation of AEAD_AES_(128|256)_GCM
may or may not be compatible with others. They are note enabled by default but may
be enabled by the user. See the Reference Manual, Application SSH for details of
how we interpret the rfc.

To be safe, use aes128-gcm@openssh.com or aes256-gcm@openssh.com instead.