Fix missing filename and line number in warning
When the 'bin_opt_info' is given, warnings without filenames
and line numbers could sometimes be produced:

  no_file: Warning: INFO: matching non-variables after
  a previous clause matching a variable will prevent delayed
  sub binary optimization

The reason for the missing information is that #c_alias{} records lack
location information. There are several ways to fix the problem. The
easiest seems to be to get the location information from the
code).

Noticed-by: José Valim