Distinguish between GC disabled by BIFs and other disabled GC
Processes remember heap fragments that are known to be fully
live due to creation in a just called BIF that yields in the
live_hf_end field. This field must not be used if we have not
disabled GC in a BIF. F_DELAY_GC has been introduced in order
to distinguish between to two different scenarios.

- F_DISABLE_GC should *only* be used by BIFs. This when
  the BIF needs to yield while preventig a GC.
- F_DELAY_GC should only be used when GC is temporarily
  disabled while the process is scheduled. A process must
  not be scheduled out while F_DELAY_GC is set.