erl_lint_SUITE: Add smoke test of format_error/1
The test suite depended on the compiler to call
erl_lint:format_error/1 to ensure that format_error/1 was
covered. Unfortunately, though, if format_error/1 crashed
the compiler would catch the exception so that the test suite
would not notice it.

Add a smoke test of format_error/1 that will crash if there is
any problem with erl_lint:format_error/1.