erts: Fix bug in heap_factory_undo for FACTORY_HEAP_FRAGS mode
Make sure a heap fragment is not deallocated before all off_heap
terms have been cleared.

The fix assumes/asserts that the off_heap-lists of all additional
heap fragments are empty.

I think this bug has been harmless as hashmap nodes, which is only
ones (?) that can cause a factory to produce more heap,
are not linked in off_heap-list.