erts: inline tag_val_def
The tag_val_def function was called and multiple switch statements
had to be traversed in term.c, and then a big switch in the calling
code to branch on the term types. By inlining the switches are
merged by the compiler and a lot fewer branches have to be taken.

Benchmarks show that this increases performance of enc_term by as
much as 10%.