Function for generating OpenSSL-style name hashes
OpenSSL has functions to generate short (eight hex digits) hashes of
issuers of certificates and CRLs.  These hashes are used by the
"c_rehash" script to populate directories of CA certificates and CRLs,
e.g. in the Apache web server.  Adding this function lets an Erlang
program find the right CRL for a given certificate in such a
directory.