observer: Optimize drawing of graphs
By allowing changes to the number of minutes displayed and update
frequency, we need to optimize the drawing of the graphs as it can
no longer recalculate everything in each frame drawn.

Only recalculate the changed entries, takes more memory but far less
cpu usage.

While updating the gui, increase the frame-rate a bit so it updates
smoother and decrease pen size for graphs to 1 pixel as it looks better
according to an office voting.