gs: Remove the contribs directory
Some of the games no longer work. For example, when calling
cols:start/0, there will be a 'badfun' exception because of
an attempt to apply a tuple fun.

There are also calls to the deprecated now/0 function and to
the deprecated 'random' module.

Since the entire gs application is deprecated and scheduled
to be removed any release now, there is no need to keep the
contribs directory.