Add erl_prim_loader:get_modules/3
When we are going to implement functions that can load many modules
at once, we don't the erl_prim_loader server to become a bottleneck.
Therefore, we need erl_prim_loader:get_modules/3 that can read many
BEAM files in parallel.

Note that we will not bother making reading from archive files
or using the inet loader efficient. That can be done later if it
turns out to be important.