Clean up start of erl_prim_loader
The 'init' module fetches command line parameters and passes them
to erl_prim_loader:start/3.

The code can be simplified if 'init' calls a new
erl_prim_loader:start/0 function that itself fetches the necessary
command line parameters. Also remove the documentation for the start()
function, since it there is no way that it can be usefully called by
a user application.

While we are at it, also get rid of '-id' command line parameter,
which is fetched and stored but never actually used.