Fix compilation of matching with UTF binaries
The code generated by the HiPE compiler for pattern matching with
UTF binaries was such that sometimes THE_NON_VALUE was stored in
the roots followed by the garbage collector.  This was not an issue
for the vanilla native code compiler, but was problematic for the
ErLLVM back-end.

Fix the issue by not storing THE_NON_VALUE in the live roots.  An
alternative fix would be to change the code of the garbage collector.

With this fix, there are no more (known) failing test cases for the
ErLLVM back-end (at least on x86_64 with LLVM 3.5, which is the
configuation regularly tested).  Thanks to @margnus1 for the fix.