beam_bool: Fix unsafe optimization
beam_bool would make the following code unsafe (which would be
reported by beam_validator):

scotland(Echo) ->
  found(case Echo of
	    Echo when true; Echo, Echo, Echo ->
		Echo;
	    echo ->
		[]
	end,
	Echo = placed).

found(_, _) -> million.

Basically, beam_bool would see that the 'case' would always return
the value of Echo. Thus:

scotland(Echo) ->
  found(Echo, Echo = placed).

The only problem is that beam_bool would also remove a 'move'
instruction that would save Echo to the stack. Here is the
assembly code for part of the function:

    {allocate_zero,1,1}.

    {move,{x,0},{y,0}}.           %% Save Echo on stack.

    {bif,'=:=',{f,7},[{x,0},{atom,true}],{x,1}}.
    {bif,'=:=',{f,7},[{x,0},{atom,true}],{x,2}}.
    {bif,'=:=',{f,7},[{x,0},{atom,true}],{x,3}}.
    {bif,'and',{f,7},[{x,2},{x,3}],{x,2}}.
    {bif,'and',{f,7},[{x,1},{x,2}],{x,1}}.
    {jump,{f,8}}.

  {label,7}.
    {move,{atom,false},{x,1}}.

  {label,8}.
    {bif,'or',{f,6},[{atom,true},{x,1}],{x,1}}.
    {test,is_eq_exact,{f,6},[{x,1},{atom,true}]}. %% Jump never taken.

    {jump,{f,5}}.

  {label,6}.
    {test,is_eq_exact,{f,9},[{x,0},{atom,echo}]}.
    {move,nil,{x,0}}.
    {jump,{f,5}}.
  {label,9}.
    {test_heap,3,0}.
    {put_tuple,2,{x,0}}.
    {put,{atom,case_clause}}.
    {put,{y,0}}.
    {line,[{location,"t.erl",5}]}.
    {call_ext,1,{extfunc,erlang,error,1}}.
    {jump,{f,5}}.

  {label,5}.
    {test,is_eq_exact,{f,12},[{atom,placed},{y,0}]}.

beam_bool would see that the is_eq_exact test at label 8 would
always succeed. It could therefore remove most of the code before
the jump to label 5. Unfortunately it also removed the essential
move of Echo to the stack:

    {allocate_zero,1,1}.

    %% Instruction incorrectly removed: {move,{x,0},{y,0}}.

    {jump,{f,5}}.

  {label,5}.
    {test,is_eq_exact,{f,12},[{atom,placed},{y,0}]}.

The root cause of the problem is that the 'move' instruction is
included in the block of 'bif' instructions before label 8.
Normally the 'move' instruction would not have been discarded,
but because the left operand to the 'or' BIF is 'true', the
entire block with 'bif' instructions are dropped.

As far as I can see, there is no gain by including 'move'
instructions in the first place. There is no way that better
code will be produced. In fact, the entire optimization can
be given up if 'move' instructions are found in the block.

Thus we can fix this bug by never including any 'move' instructions
in the block of 'bif' instructions. We can also remove all the
code that deals with 'move' instructions within blocks.

Reported-by: Thomas Arts