code_server: Tighten type of modules parameters
The following functions in the 'code' module only allow the
module argument to an atom:

  load_file/1
  load_binary/3
  ensure_loaded/1
  delete/1
  purge/1
  soft_purge/1
  get_object_code/1

Therefore, there is no reason that the corresponding implementation in
code_server should allow the module to be either an atom or a
string. Only accept an atom and remove the helper functions to_atom/1
and do_mod_call/4.