erl_prim_loader: Correct purging of the archive cache
prim_do_release_archives/3 can't make up its mind whether the
primary archive should be released or not. The key in the process
dictionary is kept, while #prim_state.primary_archive is cleared.

It seems that intent was the primary archive should be preserved,
because the function was intended to be called by a timeout routine
every sixth minute (it is not because of a bug in setting up
the timeout).

Therefore, rewrite the code to preserve the primary archive and
simplify it while at it. Also, rename prim_release_archives/1 to
prim_purge_cache/0 to make it clearer what it is doing.