epmd: support IPv6 on systems without inet_pton
If the function inet_pton is available, we use it to parse IPv6
addresses to listen on passed to us on the command line or through the
ERL_EPMD_ADDRESS variable.

If inet_pton is not available, let's not reference variables that are
only declared if it is.

This means that we won't be able to specify IPv6 addresses to listen
on on such systems, but listening on any address as well as specific
IPv4 addresses should work.